refactor(server): :recycle: refactored to result pattern in domain, a…

+++ libs/server/core/application/Features/Publishers/Commands/DeletePublisherCommandHandler.cs
@@ -1,15 +1,22 @@
-using Kathanika.Core.Domain.Exceptions;
-
 namespace Kathanika.Core.Application.Features.Publishers.Commands;
 
-internal sealed class DeletePublisherCommandHandler(IPublisherRepository publisherRepository) : IRequestHandler<DeletePublisherCommand>
+internal sealed class DeletePublisherCommandHandler(
+    IPublisherRepository publisherRepository,
+    IPublicationRepository publicationRepository
+) : IRequestHandler<DeletePublisherCommand, Result>
 {
-    public async Task Handle(DeletePublisherCommand request, CancellationToken cancellationToken)
+    public async Task<Result> Handle(DeletePublisherCommand request, CancellationToken cancellationToken)
     {
-        _ = await publisherRepository.GetByIdAsync(request.Id, cancellationToken) ??
-        throw new NotFoundWithTheIdException(typeof(Publisher), request.Id);
+        Publisher? publisher = await publisherRepository.GetByIdAsync(request.Id, cancellationToken);
+        if (publisher is null)
+            return PublisherAggregateErrors.NotFound(request.Id);
 
-        //TODO: var hasPublication = (await publicationRepository.ListAllAsync(x => x.Publisher))
+        bool hasPublication = await publicationRepository
+            .ExistsAsync(x => x.Publisher != null && x.Publisher.Id == request.Id, cancellationToken);
+
+        if (hasPublication)
+            return PublisherAggregateErrors.HasPublication;
         await publisherRepository.DeleteAsync(request.Id, cancellationToken);
+        return Result.Success();
     }
 }
